# Docs/API_Endpoints/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
...
    ProductRequest:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        price:
          type: number
          format: double
        displayName:
          type: string
        symbolIdentifier:
          type: string
        minimalWarning:
          type: integer
          format: int32
    ProductRequestSimple:
      type: object
      properties:
        price:
          type: number
          format: double
        displayName:
          type: string
        symbolIdentifier:
          type: string
        minimalWarning:
          type: integer
          format: int32
    ProductResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        price:
          type: number
          format: double
        displayName:
          type: string
        symbolIdentifier:
          type: string
        minimalWarning:
          type: integer
          format: int32
        productCount:
          type: integer
          format: int32
        subProducts:
          type: array
          items:
            $ref: "#/components/schemas/ProductResponse"
...
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
  securitySchemes:
    SecurityScheme:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication
paths:
...
  /product:
    post:
      summary: Create a Product for the currently logged-in Food Court Worker Account
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRequestSimple"
        required: true
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    $ref: "#/components/schemas/UUID"
                  price:
                    type: number
                    format: double
                  displayName:
                    type: string
                  symbolIdentifier:
                    type: string
                  minimalWarning:
                    type: integer
                    format: int32
                  productCount:
                    type: integer
                    format: int32
                  subProducts:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProductResponse"
        "400":
          description: Invalid Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Product Api
      security:
      - SecurityScheme:
        - FOOD_COURT_WORKER
  /product/admin/by_food_court_id/many/{foodCourtId}:
    post:
      summary: Create multiple Products for a given Food Court
      parameters:
      - name: foodCourtId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/ProductRequest"
        required: true
      responses:
        "201":
          description: Products created (invalid items are skipped)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductResponse"
        "400":
          description: Invalid Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Product Api
      security:
      - SecurityScheme:
        - ADMIN
  /product/admin/by_food_court_id/{foodCourtId}:
    post:
      summary: Create a Product for a given Food Court
      parameters:
      - name: foodCourtId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRequest"
        required: true
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    $ref: "#/components/schemas/UUID"
                  price:
                    type: number
                    format: double
                  displayName:
                    type: string
                  symbolIdentifier:
                    type: string
                  minimalWarning:
                    type: integer
                    format: int32
                  productCount:
                    type: integer
                    format: int32
                  subProducts:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProductResponse"
        "400":
          description: Invalid Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Product Api
      security:
      - SecurityScheme:
        - ADMIN
  /product/by_id/{id}:
    delete:
      summary: Delete a Product by its ID
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: Product deleted
          content:
            text/plain: {}
        "400":
          description: Invalid Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Product Api
      security:
      - SecurityScheme:
        - FOOD_COURT_WORKER
        - ADMIN
  /product/list:
    get:
      summary: List Products for the currently logged-in Food Court Worker Account
      responses:
        "200":
          description: Products found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Product Api
      security:
      - SecurityScheme:
        - FOOD_COURT_WORKER
  /product/list/by_food_court_id/{foodCourtId}:
    get:
      summary: List Products by Food Court ID
      parameters:
      - name: foodCourtId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: Products found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductResponse"
        "400":
          description: Invalid Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Product Api
      security:
      - SecurityScheme:
        - GUEST
  /product/list_all:
    get:
      summary: List all Products
      responses:
        "200":
          description: Products found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Product Api
      security:
      - SecurityScheme:
        - GUEST
        - FOOD_COURT_WORKER
        - ADMIN
  /product/{id}:
    get:
      summary: Get a Product by ID
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    $ref: "#/components/schemas/UUID"
                  price:
                    type: number
                    format: double
                  displayName:
                    type: string
                  symbolIdentifier:
                    type: string
                  minimalWarning:
                    type: integer
                    format: int32
                  productCount:
                    type: integer
                    format: int32
                  subProducts:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProductResponse"
        "400":
          description: Invalid Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Product Api
      security:
      - SecurityScheme:
        - GUEST
...
info:
  title: App API
  version: 0.1.0-SNAPSHOT